uievents/order-of-events/focus-events/focus-contained-manual.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8">
  <title>Focus-related events should fire in the correct order</title>
  <link rel="help" href="https://w3c.github.io/uievents/#events-focusevent-event-order">
  <meta name="flags" content="interact">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/resources/event_order.js"></script>
</head>

<body>
  <ol>
    <li>Click into the text input.</li>
    <li>Click just below the text input.</li>
    <li>Click into the text input again.</li>
    <li>Click the "Done" button.</li>
  </ol>
  <div id="a" tabindex="1" style="width:400px; height:300px">
    <br /><br /><br />
    <input type="text", id="b"></div>
  </div>
  <button type="button" id="done">Done</button>
</body>

<script>
setup({explicit_timeout: true});
var events = [];
var targets = [];
function record(e) {
  e.stopPropagation();
  events.push(e.type);
  targets.push(e.target.id);
}
var relevantEvents = [
  'focus',
  'blur',
  'focusin',
  'focusout'
];
window.onload = function () {
  var a = document.getElementById('a');
  var b = document.getElementById('b');
  var inputs = [a, b];

  for (var i = 0; i < inputs.length; i++) {
    for (var k = 0; k < relevantEvents.length; k++) {
      inputs[i].addEventListener(relevantEvents[k], record, false);
    }
  }

  async_test(function(t) {
    document.getElementById('done').addEventListener('click', function () {
      var eventOrder = ['focusin', 'focus', 'focusout', 'focusin', 'blur', 'focus',
        'focusout', 'focusin', 'blur', 'focus', 'focusout', 'blur'];
      var targetOrder = ['a', 'a', 'a', 'b', 'a', 'b',
        'b', 'a', 'b', 'a', 'a', 'a'];
      t.step(function () {
        assert_true(checkOrder(events, targets, eventOrder, targetOrder));
        t.done();
      });
    }, false);
  }, 'Focus-related events should fire in the correct order');
};
</script>
</html>
